+++ cloud-scanner-cli/src/main.rs
@@ -4,9 +4,6 @@ use clap::{Parser, Subcommand};
 extern crate log;
 extern crate loggerv;
 
-#[macro_use]
-extern crate rocket;
-
 #[derive(Parser, Debug)]
 #[clap(author, version, about)]
 /// List aws instances and their environmental impact (from Boavizta API)
@@ -111,39 +108,7 @@ async fn main() -> Result<()> {
         SubCommand::ListInstances {} => {
             cloud_scanner_cli::show_instances(&args.filter_tags, &region).await?
         }
-        SubCommand::ServeMetrics {} => {
-            run_metric_server().await?;
-        }
+        SubCommand::ServeMetrics {} => cloud_scanner_cli::expose_metrics(&api_url).await?,
     }
-    Ok(())
-}
-
-#[get("/")]
-fn index() -> &'static str {
-    "Hello, world!"
-}
-
-#[get("/metrics")]
-async fn metrics() -> String {
-    let hours_use_time: f32 = 1.0;
-    let tags = Vec::new();
-    let aws_region = "eu-west-1".to_string();
-    let api_url = "https://api.boavizta.org";
-    let metrics = cloud_scanner_cli::get_default_impacts_as_metrics(
-        &hours_use_time,
-        &tags,
-        &aws_region,
-        api_url,
-    )
-    .await;
-    metrics.unwrap()
-}
-
-async fn run_metric_server() -> Result<(), rocket::Error> {
-    let _rocket = rocket::build()
-        .mount("/", routes![index, metrics])
-        .launch()
-        .await?;
-
     Ok(())
 }
